+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `erc_warnings` pre-flight option to consider ERC warnings as errors.
 - Pattern expansion in the `dir` option for outputs (#58)
+- New filter type `suparts`. Adds support for KiCost's subparts feature.
 
 ### Changed
 - Errors and warnings from KiAuto now are printed as errors and warnings.